Use img tool in pull repo rule
This change adds a new option to the "pull" repository rule: Users can now choose between using the Bazel downloader or the img tool when pulling base images.
The img tool is generally faster and more standards compliant as it uses go-containerregistry to implement authentication and the container registry API spec.