pull: handle missing "registry" attr correctly (#72)
We need to record what registry an image comes from to support shallow base images.
When no registry is specified, pulling works, but the registry wasn't recorded.
This leads to issues at deploy time, where the layers of the base image
need to be streamed in if they are missing.
Without a proper source, it's unlcear where to stream them from.

This change ensures we always record a registry for any dowloaded image,
even if the registry attr is left uninitialized.

Fixes #65