tar layer creation: calculate digests of inputs in advance (#73)
This change adds worker goroutines that inspect inputs files used to create a layer.
The inputs will be hashed in advance to prefill the cache used when deduplicating files while building the tar file.
This change should help to improve layer creation, as the throughput is currently limited by how fast we can feed data into the compressor.