push/load/multi_deploy: correctly choose img tool for host (#86)
The following use case was broken accidentally:

```
bazel run --platforms=//platforms:linux_arm64 //app:my_image
```

In this case, we want to build the image for the target platform (Linux, arm64) but still need to choose a push binary that can run on the host platform (macOS, Windows, different arch, ...).

The `host_platform_transition` was supposed to always transition to the host platform but didn't.